Lifted state up to Discover class component to hold data for ProfileCard and ProfileDetail class components.
