+++ client/app.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Home from './pages/home';
+import Discover from './pages/home';
 
 export default class App extends React.Component {
   render() {
-    return <Home />;
+    return <Discover />;
   }
 }
