+++ client/components/logo-navbar.jsx
@@ -8,9 +8,9 @@ export default function LogoNavbar(props) {
         <a className='navbar-brand ms-4 me-4 pt-0 pb-0' href="#">Woofles</a>
         <i className="fa-solid fa-stroopwafel me-5"></i>
         <div id='desktop-nav'>
-          <a href="#" className='anchor ms-3'>Discover</a>
+          <a href="#discover" className='anchor ms-3'>Discover</a>
           <a href="#favorites" className='anchor'>Favorites</a>
-          <a href="#Inquiries" className='anchor'>Inquiries</a>
+          <a href="#inquiries" className='anchor'>Inquiries</a>
         </div>
       </div>
     </nav>
